refactor tenant
